move getAttackTimings to core to avoid core importing a public module
